+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+### 1.4.6
+- Fix issue with plugin breaking after parsing a file containing theorem without `Proof.` keyword.
+
+### 1.4.5
+- Fix formatting issues when inserting the proof into the editor.
+
+### 1.4.4
+- Do not require a theorem to be `Admitted.` for coqpilot to prove holes in it.
+- Correctly parse theorems that are declared with `Definition` keyword.
+
+### 1.4.3
+- Tiny patch with shuffling of the hole array.
+
 ### 1.4.2
 - Now no need to add dot in the end of the tactic, when configuring a single tactic solver.
 - Automatic reload settings on change in the settings file. Not all settings are reloaded automatically,